Re-factors the hamlib and mode/band changing of the receiver.
Don't think I've broken too much in the process but some testing will probably
show up some bad things.

Should now (or when the bugs are found) remember the sound card and rig in use
for each mode/band and change back to it when the mode/band are changed.

Significant changes to Dream.ini too so best to start clean.