progress towards a command line with Qt operation.
This can now be built with

qmake CONFIG+=console dream.pro
make

It builds OK on linux with Qt 4.6.

Console app depends on Qt Core, Xml and Network. GPSd is not supported as its invoked from the GUI. It could easily
be given its own thread.

main now supports the --help invocation and runs the DRMReceiver instance in a thread so that the RSCI input can work
as the latter uses QTimer for file input and Socket signals for IP input. It seems to run but file the OnDataReceived slot of the PacketSourceFile is not working. I've left some print statements in to help.