Committing this but it is really just some ideas. AudioDetailWidget does not play well with engineering tabs in a vertically separate tabwidget - we don't need two charts at once here. Auto-resizing is both nice and nasty. I think it might be good if it could be animated so the transitions were smooth.
Ideas on how to arrange the engineering views of services and the
channel/stream etc. widgets would be welcome.

Still missing from the single window mode are the stations and live
schedule windows. Should these be in the service tab widget? It isn't
that easy to keep the ordering sane. Again, other peoples ideas are
welcome.


git-svn-id: https://svn.code.sf.net/p/drm/code/dream@853 ddb839e9-6489-4ed3-a1ee-f9006aeb25b6